♻️ 🚚 Refactor quintly client requiring
