+++ okr/scrapers/quintly.py
@@ -3,25 +3,28 @@ import datetime
 
 import numpy as np
 
-from analytics import quintly
+from analytics.quintly import QuintlyAPI
 
 
-_quintly = None
+quintly = None
 
 
-def get_quintly():
-    # Instantiate the class with your client id and secret
-    global _quintly
-    if not _quintly:
-        _quintly = quintly.QuintlyAPI(
-            os.environ.get("QUINTLY_CLIENT_ID"), os.environ.get("QUINTLY_CLIENT_SECRET")
-        )
-    return _quintly
+def requires_quintly(func):
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        global quintly
+        if not quintly:
+            quintly = QuintlyAPI(
+                os.environ.get("QUINTLY_CLIENT_ID"),
+                os.environ.get("QUINTLY_CLIENT_SECRET"),
+            )
+        return func(*args, **kwargs)
+
+    return wrapper
 
 
-# You can run the query with the run_query method. It returns a pandas DataFrame
+@requires_quintly
 def get_insta_insights(profile_id, *, interval="daily", start_date=None):
-    quintly = get_quintly()
     profile_ids = [profile_id]
 
     if start_date is None:
@@ -71,8 +74,8 @@ def get_insta_insights(profile_id, *, interval="daily", start_date=None):
     return df
 
 
+@requires_quintly
 def get_insta_stories(profile_id, *, start_date=None):
-    quintly = get_quintly()
     profile_ids = [profile_id]
     table = "instagramInsightsStories"
     fields = [
@@ -96,8 +99,8 @@ def get_insta_stories(profile_id, *, start_date=None):
     return df
 
 
+@requires_quintly
 def get_insta_posts(profile_id, *, start_date=None):
-    quintly = get_quintly()
     profile_ids = [profile_id]
     table = "instagramOwnPosts"
     fields = [
@@ -135,8 +138,8 @@ def get_insta_posts(profile_id, *, start_date=None):
     return df
 
 
+@requires_quintly
 def get_youtube_analytics(profile_id, *, interval="daily", start_date=None):
-    quintly = get_quintly()
     profile_ids = [profile_id]
     table = "youtubeAnalytics"
