+++ okr/views/insta.py
@@ -7,7 +7,7 @@ from django.views.decorators.http import require_POST
 from django.views.decorators.csrf import csrf_exempt
 
 from ..models.insta import *
-from .. import quintly
+from ..scrapers import quintly
 
 berlin = timezone("Europe/Berlin")
